refactored application
